+++ src/4 - Domain/Models/Domain.Models/Dto/TotalPriceTypeActivesDto.cs
@@ -5,7 +5,7 @@ namespace Service.Core.Dtos
     public class TotalPriceTypeActivesDto
     {
         public EnumCategory TypeActives { get; set; }
-        public double TotalCost { get; set; }
-        public double TotalCostPercent { get; set; }
+        public decimal TotalCost { get; set; }
+        public decimal TotalCostPercent { get; set; }
     }
 }
